feat($runner): implements simple reporting
The Runner class now measure the loading and excution duration
of specs and offer a results summary at the end of the output.
Failed specs are now badged with red FAIL prefix.

BREAKING CHANGES:
Skipped specs are now considered as failures. Skipped specs
are used whenever a specs dependency is not met, even if the
skip helper is provided for test purpose. Leaving hardcoded
skip call or having unmet dependencies in test have to be
considered as errors and reported.