feat(factories): implements instance method descripbe
Implements the handling of string such as '::method' in the describe
function. In that case, the actual subject is used as the constructor
for the instance own the described method. The instance is created
using the build function provided by the factories module.
The instanciation use the parameters defined with the withParameters
helper when provided.