feat(matchers): add matchers loading from support
The matchers are automatically loaded from the specs/support/matchers
directory unless the --no-matchers option is passed to spectacular.
The path in which looking for matchers can be changed using the
--matchers option.
As a side effect, all the instructions in the index was moved into
the run block to allow to access options passed by the binary.