feat(formatters): add a resume with errors, failures, skipped and pending
When a lot of failures or errors have been found the log can be quite
hard to read. The resume at the end will allow to have a quick summary
of the results with index for each categories's entry.
As a reminder, the skipped and pending tests are listed here as well.