refactor(matchers): better find spec file in stack
Since matchers are asynchronous the stack trace can be a little
messed. In order to find the file that defined the expectation we
need a method that compare a stack with the specs paths found by the
runner. If a spec file is found in the stack the stack is cropped up
to this point.