fix(matchers): promise based matchers's stack trace
Promise are executed in a `setImmediate` callback, and doesn't
know about the expectation creation stack. In that case it's
not possible to find the spec file in an error's stack.
To prevent that, an error is created on a should call and store
in the expectation. If the matcher end either with true or false
(no error was raised, no timeout), the passed-in error is used
to find the spec file.