wip(server): introduce a server option for running specs in browser
In order to run specs in a browser the build has been changed to join
all files loaded into a vm inside a spectacular.js file.
A server option was added that, when set, will force the spectacular
bin to run a server rather than running the specs on node.