refactor($runner): switch to an event based logging
The ConsoleReporter no longer have node dependencies. It is registered
to the Runner events and itselsf dispatch message and report events.
The result formatting process is now mostly asynchronous and promise
based. It allow to have the file's loading performed even in browser
(where file's loading is asynchronous).
The environment no longer stores the formatter class.