feat(fixtures): implements fixture function
The fixture function is part of the environment in order to use
data provided in the options.
Files are loaded asynchronously and according to the context (node,
browser) thanks to the loadFile option that is already used by
the reporters.
Files processing after loading is also done asynchronously using
a promise.
Custom handler can be added to an environment using the
Environment::registerFixtureHandler