refactor(reporters): browser stack reporter now extends the console one
There was too much duplication in these two classes and their names
was too ambiguous. Now the console stack reporter is named StackReporter
and the browser one BrowserStackReporter.
The file's loading and stack formatting is now completely handled
by the StackReporter.
The BrowserStackReporter now only handle the display part of the process.