feat(mixins): implements a Globalizable mixin
The Globalizable mixin provides functionality to an instance to
decorate the global object with its own methods or properties.
Methods are bound to the instance that was loaded.
The globalizable mixin take care to keep a reference to the previous
values in the global object and to restore them on a unglobalize call.
The Trait and Factory class are now Globalizable.