refactor(environment): add methods for aliasing globalizable methods
The Environment now provides four methods to alias either it, describe,
or other methods that have to run inside or outside example blocks.
These alias are then handled with the createGlobalizedMethod that will
take care to append the method to the global object if the current object
is globalized.
Alias already existing for Environment's methods are now created
with these helpers.