feat(environment): add expect(...).to and expect(...).not.to syntax
The expect method accept either a value or a description and a value and
returns an object with a to method and a not sub object that also contains
a to method.
It creates an expectations for the passed-in value with either the
value or the description in the expectation message.
The Expectation class now accept a description that will be prepended
to the matcher description.