feat($core): remove dependency to jquery for dom specs
The DOMExpression class and the match and contains matcher doesn't
rely on jquery anymore. Instead the DOM methods are used (matchesSelector,
querySelector, querySelectorAll).
The tests that was using a jquery object was refactored to use the
documents methods.
The HTML fixture handler now also use the documents methods to inject the
fixtures in the DOM.