feat(Cakefile): add browser packaging task
The cake build task compiles the files, join them, produce a minified
version, generates the stylesheet and zip the files.

The tests are compiled as well to allow to run them in the documentation.

To support all these operations and ease the work with async commands
most of the work is done with promises.