feat(factories): implements factory inheritance
By using the extends option instead of the class option you can specify
a parent factory for the current factory. All the traits and setup of
the parent factory can be used. The current factory setup takes priority
on the parent one when a conflict occurs.
Reopening a factory doesn't imply to pass an option object, using only
the name is enough.
Missing trait and missing parent now raises an exception.