fix(examples): passing `undefined` to `describe` is now a valid subject
If a test using `itsInstance` was dependent of its parent and that parent
group has was called with `undefined` as subject it was failing
in `itsInstance` instead of being skipped due to the parent tests failing
due to the `undefined` subject.

By not setting a default value for the `desc` argument it allow
the constructor switch to handle undefined values as legitimate
subjects.

The root group is now initiatlized with an empty string to force the
lack of subject and allow `itsReturn` and `itsInstance` to work properly.