refactor(options): group contextual methods definition in a function
In order to make it more clear what is injected in the options given
the context (node or browsers), the `loadFile` and `getOriginalSourceFor`
methods are now injected by the CliMethods function.
The `getOriginalSourceFor` also rely on a `hasSourceMap` method that
return true for files that have a source map available.