feat(source-map): implements source map support in the browser
The BrowserMethods decorates the options with the contextual methods.
Two extras methods are provided to allow to define where the source
and source map files can be found when a file returned true in
`hasSourceMap`.
The server now provides routes to return source and source map and
specify the url mapping to use to retrieve these files.